refactor: extract map runner policy metadata helpers
